Make type float instead of deciaml to easy deserialize

+++ app/energy_agent/energy_decrypter.py
@@ -342,7 +342,7 @@ def transform_to_metrics(data_list, public_key) -> dict:
     }
 
     for data in data_list:
-        value = Decimal(data.get("value"))
+        value = float(data.get("value"))
         if data.get("key") == "WirkenergieP":
             metric_data["absolute_energy_in"] = convert_to_kwh(value)
         elif data.get("key") == "WirkenergieN":
